Some more changes.

+++ client_comm.h
@@ -2,7 +2,7 @@
 #ifndef CLIENT_COMM_H_
 #define	CLIENT_COMM_H_
 
-int get_connected_socket(char * server_address, int server_port);
+int get_connected_socket(char * server_address, unsigned short server_port);
 int run_client(char *	server_address, int server_port,
 char * username, char * password);
 int process_server_request(int fd);
